+++ openstack/db/v1/configurations/testing/requests_test.go
@@ -235,21 +235,3 @@ func TestGetGlobalParam(t *testing.T) {
 
 	th.AssertDeepEquals(t, expected, param)
 }
-
-func TestAttachConfigGroup(t *testing.T) {
-	th.SetupHTTP()
-	defer th.TeardownHTTP()
-	fixture.SetupHandler(t, resURL, "PUT", "", "", 202)
-
-	_, err := configurations.AttachConfigGroup(fake.ServiceClient(), "0000-0000-0000-0000", configID).Extract()
-	th.AssertNoErr(t, err)
-}
-
-func TestDetachConfigGroup(t *testing.T) {
-	th.SetupHTTP()
-	defer th.TeardownHTTP()
-	fixture.SetupHandler(t, resURL, "PUT", "", "", 202)
-
-	_, err := configurations.DetachConfigGroup(fake.ServiceClient(), configID).Extract()
-	th.AssertNoErr(t, err)
-}
